[LB-457] feat : 트레이싱 설정
