+++ src/main/java/com/lgcms/core/config/KafkaConfig.java
@@ -1,0 +1,9 @@
+package com.lgcms.core.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.annotation.EnableKafka;
+
+@Configuration
+@EnableKafka
+public class KafkaConfig {
+}
